+++ BREAKING_CHANGES_LOG.md
@@ -3,6 +3,15 @@ Before 1.0, the stack do NOT follow semver version in releases.
 This document aims to ease the WIP migration from a version to another by providing intels about what to do to migrate.
 
 ## v0.156.0
+* cmf: selectors
+* PR: https://github.com/Talend/ui/pull/1055
+* Change: move to collections
+
+| name | new location |
+|---|---|
+| getCollectionFromPath | selectors.collections.find
+| findCollectionPathListItem | selectors.collections.findListItem
+
 * cmf: putActionCreator
 * PR: https://github.com/Talend/ui/pull/1055
 * Change: move from api.saga.putActionCreator to api.sagas.putActionCreator
