+++ packages/cmf/src/sagas/component.js
@@ -1,7 +1,29 @@
 import invariant from 'invariant';
 import { fork, cancel, take, takeEvery } from 'redux-saga/effects';
 import CONST from '../constant';
-import { get } from '../saga';
+import registry from '../registry';
+
+/**
+ * This function register a saga in the cmf registry
+ * @param {string} id the saga id you want
+ * @param {generator} saga the saga generator
+ * @param {object} context optional context to get the registry
+ */
+export function register(id, saga, context) {
+	registry.addToRegistry(`SAGA:${id}`, saga, context);
+}
+
+/**
+ * This function allow to get a saga from the registry
+ * @param {string} id the saga id you want
+ * @param {object} context optional context to get the registry
+ */
+export function get(id, context) {
+	return registry.getFromRegistry(`SAGA:${id}`, context);
+}
+
+export const registerMany = registry.getRegisterMany(register);
+
 
 export function* onSagaStart(action) {
 	const saga = get(action.saga);
