+++ packages/cmf/src/sagas/index.js
@@ -4,33 +4,12 @@ import putActionCreator from './putActionCreator';
 import * as collection from './collection';
 import * as component from './component';
 
-import registry from '../registry';
-
-/**
- * This function register a saga in the cmf registry
- * @param {string} id the saga id you want
- * @param {generator} saga the saga generator
- * @param {object} context optional context to get the registry
- */
-export function register(id, saga, context) {
-	registry.addToRegistry(`SAGA:${id}`, saga, context);
-}
-
-/**
- * This function allow to get a saga from the registry
- * @param {string} id the saga id you want
- * @param {object} context optional context to get the registry
- */
-export function get(id, context) {
-	return registry.getFromRegistry(`SAGA:${id}`, context);
-}
-
-export const registerMany = registry.getRegisterMany(register);
-
 export default {
-	get,
-	register,
-	registerMany,
+	// shortcut
+	get: component.get,
+	register: component.register,
+	registerMany: component.registerMany,
+
 	changeDocumentTitle,
 	collection,
 	component,
